Initialize PL-lucia branch
In addition to creating the branch for working on running Lucia within
the rest of Rosie, some changes were needed to get it to work.  The big
one is that WIndows will not allow a file whose name is "aux", so
changed aux.soar and references to it to aux1.soar.  Now Windows is
happy.  Also tweaked the _agent.soaragent file to get rid if an problem
in Eclipse.