First steps to integrate Lucia
Several changes made on the PL-lucia branch, which was just branched off
of development.
A debug parameter of debug-wait-limit was added and code added to
wait.soar to interrupt after the given number of waits in a
state-no-change.  Works as always if this is 0.
Added and agent-param called language-mode which can be set to laird or
lucia.  Right now it only controls what world to initialize to.
Added a lucia folder with a config file, a source file,
initialize-lucia.soar, and a simple script.
Set up _agent_source to it sources both language-comprehension and lucia
right now.  Since there's no real substance in lucia yet, the system
runs using the laird parser regardless of the language-mode, which only
affects what world to use for now.
Added a line to BasicRosieGui.java to make it read the script file
specified in the config file.  Removed some unused imports in two java
files.
So now it runs with John's parser but the Lucia's world.