Setup rosie on PL laptop
Needed to make some adjustments to get the rosie project to run properly
on PL's laptop.  The first thing was the working directory.  Set that in
the run configuration to "../rosie" and then modified the paths in the
lucia.config file to work from that base folder.
Also had to recompile the latest 9.6.0 Soar, and get the correct
SOAR_HOME to point to this version of Soar.  It appears that setting
SOAR_HOME in an Eclipse run configuration does not work, it has to be
set in the Windows Environment Variables.  With all this working, saved
a launch configuration in the project for future reference.
Hopefully all this will work on either the PL laptop machine or on
Xiang-liang in the Soar lab, but that remains to be confirmed.