Got messages to Rosie!!
Added code to simulate comprehension of "Pick up the green sphere." and
"Put it on the stove." in lucia mode.  The Rosie version of the
comprehend operator runs under the dialog-event state and goes to a
substate.  In that substate there is no code to do anything, so in
another impasse simulate-comprehension runs and posts a simulated
message to the dialog-event state.  Next a simulation version of
interpret-for-rosie runs and puts the message on the top state's
interaction status along with the dialog-event that has been performed.
Now Rosie takes over and processes the message.
It appears the message processing is not really complete yet because the
action has not been properly grounded.  Not sure if the object grounding
works with Rosie yet.