Start merge with Lucia
The goal is to get all the Lucia comprehension code working in the rosie
project under the lucia directory.  As much as possible, we would like
all these files to be identical.
So far all the code under comprehend was copied over from the Lucia
project., and comprehend.soar was copied and made it work in both
environments.  Some updates were also made to Lucia_source.soar and
initialize-lucia.soar.
Then comprehend.soar was worked over so it has two versions of the
proposal, one for each system, and this same identical file works in
both systems.
Much work remains to be done.