Added the Rosie grammar
Copied both the ecg and soar directories for the Rosie grammar in Lucia
to a new Grammar directory under lucia in the rosie agent.  With those
files there, and a change so that they get sourced into the agent, the
agent parses "Pick up the green sphere." correctly!!
There are still some issues to address:
1. In rosie there is an extra '.' word at the end.
2. The numbers under ^source end up as 1-based here, whereas they are
0-based in Lucia.
3. We get no reports of constructions or meanings on the console.
4. There are still 8 files that don't match between the two systems.
All these issues will be dealt with in due time.