Got 2 sentences working with Rosie
This is being committed on Windows, so it has not been tested with the
actual arm simulator.  But Lucia is generating messages that Rosie
processes properly for "Pick up the green sphere." and "Put it on the
stove."
All the files in the Lucia interpret directory were copied into the
rosie/lucia/interpret-for-rosie directory, and the problem-space was
changed from interpret to interpret-for-rosie throughout.  A few of the
cases in do-action were changed to produce Rosie messages directly with
no need for an act operator.  However, the rest of the cases in
do-action and all cases in the rest of the files remain to be changed.