Got 2 test S's to run with both parsers
Several changes were made to make this work with the test-agents/lucia
data.
1. The world for lucia is now initialized through the configuration from
the world1.txt file using the create-internal-world operator instead of
in the initialize-lucia operator.
2. The old version of the world was renamed world0.txt and the world
data that conforms to all my 209 testing is in world1.txt.
3. The "1" at the end of name definitions in smem.txt was removed since
names should be in English.  The world and lucia's grammar conform to
this.
4. A ConfigureAgent run configuration was built to run all the
configuration from Eclipse.
5. BasicRosieGUI (1).launch was created to run the system with the two
sentences in sentences.txt.
6. BasicRosieGUI.launch was edited to run in this context.  It should
get its sentences from the GUI, but that is not working for some reason.
7. Now both parsers run correctly for these two sentences!  The agent
even claims to do the correct actions.  The Soar traces are saved in
laird-first-2.txt and lucia-first-2.txt.
The parser can be chosen by an option in the sim_tabletop.config file,
which is left set to lucia.