Added the parser-test option and made it work
This change produces a system that can run either parser in test mode.
It works for at least two test sentences in four modes, laird or lucia
and test or not.  The changes are:
1. Change the RosieConfig.java to have a new option called "parser-test"
which can be either "true" or "false".
2. Put this option in Lucia's sim_tabletop.config.
3. Add code to the RosieAgentConfigurator.java to source the
test-parser.soar file, which has a bunch of rules for testing.
4. In the test-parser code, put a number of rules to set up sentences
from the script and increment through them, parsing one at a time.
5. Also in the test-parser code add rules to replace the normal
proposals for Lucia's comprehend and interpret operators and Laird's
comprehension operator so that these operators run on the top state.
6. Added code to test-parser to post the messages produced by each parse
to the testing data.
7. Added a line to several proposal rules to inhibit them from running
in test mode so that the whole interaction logic does not run in this
mode.
8. Captured and saved log files for each of the four modes to remember
how they worked at this stage.