Interpretation working in Rosie/Lucia test
This commit has gotten interpretation working, at least for two test sentences, in Rosie using a Rosie version of the interpret and act operator proposals along with all the code for applying those operators directly from the Lucia code base.  Thus the whole interpret-for-rosie operator and its implementation are being deleted.
This requires a new version of the proposals for interpret and act, but all the sub-operators that implement them are kept intact, in fact this version just loads them directly from the Lucia project code.
In addition, some things are added to make the Lucia logic work in the Rosie environment.  The sentence-complete operator runs after act, deletes all the Lucia-specific things from the top state, and puts the final action message on Rosie's interaction stack.
The test-parser.soar file has been changed to work correctly.  It now has several operators: test-first-sentence, test-next-sentence, update-test, and script-done.  All these are used for both the laird and lucia parsers.
A bug was fixed in RosieAgentConfigurator.java so that it now outputs the parser-test parameter to the file to set agent-params.
The ConfigureAgent.launch file has been changed to work on PL-Laptop instead of my UoM desktop.