Lucia working in Rosie!
This commit fixes a couple of problems that prevented Lucia from working properly in the full Rosie agent, with interaction and everything.
First, one line was removed from the rosie-mode proposal for act.  This had prevented it from firing in the dialog-event state.
Second, several changes were made to sentence-complete.  Now there are two full versions, one for test mode and one for normal mode.  Also, additional code was needed in the apply rule for the normal mode to get it to put a "performed" object on the interaction.status.  Without this, the interaction logic would not work properly.
Logs of running both parsers in normal mode with a two-sentence script are saved under agent/lucia.
There are minor differences in the output messages from the two parsers for the two sentences tested here.  For "Pick up the green sphere." laird has an extra ^modifier N7 on the command, with (N7 ^handle upwards).  This may not be needed, and lucia does not produce it.  For "Put it on the stove." laird has no arg1 on the command, but lucia has ^arg1 O20, where O20 is the green sphere that was picked up.  These differences are probably ignored by the SentenceTester, but it's not clear if they might affect the operataion of the agent.
For the first stentenced, both parsers give this action:
!ACTION! op_pick-up1 object-12
Looking at the execution part of the logs, the modifier in the first sentence appears to make no difference for the exection.
The execution of the action for the second sentence, however, is quite different.  For the second sentence, the actions are:
laird: !ACTION! op_put-down1 object-1
lucia: !ACTION! op_put-down1 object-12 object-1
Clearly the extra argument messes up the execution.  The laird execution does a bunch of stuff that the lucia one does not.  The lucia one appears to get confused and give up.  This will need to be fixed at some point, along with other important differences that may be ignored by the SentenceTester.