Got some tests running, still bugs
Added some logic to various parts of the rosie project so that there is now a good way to run tests on both parsers and get logs out that could be used by the SentenceTester.
This included making new config files and launch configurations to run the two tests, and some changes to the configuration tool to make this work.
When running the tests using the Basid50.txt input file, both parsers failed partway through: lucia on #13 and laird on #16.  There are also some minot formatting issues with the log files produced.
The testing method still needs to be documented somehow.
With these limitations, still committing this much.