Ran Lucia test on 209ordered.txt - works +/-
One problem with running Lucia in Rosie for the 209ordered.txt file was that sentence #61 "What is inside the pantry" has no terminating punctuation.  The program that formats sentences for input to Rosie for testing combines #61 with #62 since it goes until it sees a terminator.  At some point it would be nice to fix this, but for now a '.' was added to the end of that sentence and now it works.
A second problem was that next-word.soar was not looking for any terminators other than '.'.  This was changed, and now there are four different CASEs for getting the next word in Rosie.
A small bug in test-parser.soar was fixed so it keeps the sentence count correctly.
The config files were changed to make it easy to switch input files.
With all these changes made, the test of Lucia in Rosie on parser-test mode was run on 209ordered.txt.  There were no crashes.  However, 114 of the sentences had some kind of FAILURE, and a version of the log that shows this was saved away in lucia-log-FAILUREs.txt.  Analysis showed that about 20 of these failures are due to differences in the handle or name used to lookup items in smem.  Most of the rest are due to words that appear to be unknown in the grammar, with a couple of syntax errors.
It seems likely that these errors are due to an old version of the grammar, but no fixes were made for this commit to avoid confusing too many issues.
For some reason the latest run has only 111 FAILUREs.  That is the version that was saved for further work.