Updated grammar to match Lucia
The grammar under rosie/agent/lucia was a year old and had not been updated.  Running with that gave 111 FAILUREs, most of which were "Unknown-word" issues because lots of newer vocabulary was not being loaded.
Replaced all the grammar files from the Lucia project, and changed load-grammar.soar to load all the newer files that were not there before.
Now there are 114 errors, but they are all semantic errors with only a single unknown word, the one that is exprected.  These semantic errors seem to be due to differences in the names or handles used to lookup things in smem, which seem to be different from the ones we used in the Lucia project.  More work will be needed to fix all these.
The lucia-log-FAILUREs.txt file was updated to reflect the latest run so that we can use this as a reference while working on the sem errors.