Align Lucia with Rosie #1
There is a commit of this same name in the Lucia project.
This set of changes basically makes a number of vocabulary items work in Rosie that did not before by changing Lucia's grammar and smem to have the same handles/names for those items as Rosie, and than copying those grammar changes into the Rosie project.
After the last commit there were 114 FAILUREs in lucia-log.txt.  Now there are only 58.  Those that remain include 27 ^action, 15 ^2 and 12 ^arg1 errors, plus a few other miscellaneous.  There are now still 19 "syntactic" errors, which overlap with the others.
Still more to come.