Added lucia-extras to smem
A number of FAILUREs in testing 209ordered.txt in Rosie with Lucia were due to information in Lucia's smem that was not there in Rosie's smem.  Thus a new file called lucia-extras.soar was added to the agent/init-smem directory, and smem.txt for the Lucia test agent had a line added to include this file in the configuration.  Now the number of FAILUREs has been reduced from 58 to 30.  Many or those that remain are syntactic.
This change does not involve any corresponding change in the Lucia project.