Align Lucia with Rosie #2
There is a commit with this same name in the rosie project.
On the Lucia side the handle for "heard" was changed from "heard" to "heard1" in both the grammar and smem.  Both comprehend-done and next-word were edited so that the same file works in both systems using some different apply rules.
On the Rosie side, several new concepts were added to lucia-extras, the changed xv-verbs files were copied in from Lucia, and Lucia_source.soar was changed to get the entire version A of comprehend from Lucia.  This last included the combined versions of comprehend-done and next-word.
Now Rosie runs the 209ordered.txt file in Lucia test mode with only on FAILURE, which is the expected one for "unkown."
How ever, there are two remaining issues which remain to be resolved:
On #86 the system stops for long periods on the interpret and do-action operators.
On #119 11 results with sequential numbers are printed out, sometimes.
Also, the code for sourcing from Lucia should be cleaned up somehow.