Tested Laird parser on the 209
Tried running the laird parser on 209ordered.txt.  Found 30 missing vocabulary words, which are listed in "Laird unknown-words.txt."  Still 13 sentences crashed the parser, and they are listed in "Laird failed sentences.txt."  Some caused an interrupt, and others died in an infinite loop.  A copy of the original list with the 13 failed sentences commented out is in "209ordered - failed.txt."
Also changed ROSIE-HOME in the configuration launch files to work on my UMich desktop.