Starting to test eight-puzzle
Set up the lucia test agent with the data needed to test the script for eight-puzzle.  Also added a debugging option to interrupt after each sentence in parser-test mode.
At this point the laird parser does reasonably well in test mode.  In rosie mode it crashes in trying to execute the action defined by the first sentence.
Lucia has lots of missing grammar and vocabulary.  Some of this has been implemented in the Lucia project, but none has been ported to Rosie yet.  In test mode the parser crashes somewhere in ground-this-n-that.