Add "hypothetical" parameter
Added a parameter called "hypothetical" to the config files and their processing program.  This gets turned into a flag on agent-params on the top state.  Added an elaboration to the laird parser that sets this flag on a segment if both the parser-test and hypothetical parameters are set to true.  This allows the laird parser to run properly for game scripts in parser-test mode.
It would also be nice to turn off all the ^referent-info stuff in this mode, since it's output is enormous.