Added a DEBUG option to RosieCLI
These changes make it so if you put "DEBUG" as args[1] when running RosieCLI it will start the SoarJavaDebugger in a clean way as a single window, and the whole thread exits when that window is closed.  Without this option another simple window is also created, and when the windows are closed the thread keeps running.