Set up for multiple config files
To do the testing we need 2x2x2 different config files: 2 parsers, test or rosie mode, and games or robot.  Reorganized some of the configuration stuff to make that possible.  Arranged a generic.config file that has all the options and can be customized.  Made one custom version so far: lucia-test-robot.config and got it to work.  In that mode the parser runs as expected, but it needs a lot of work to comprehend all the sentences correctly.