Testing laird parser with 209ordered sentences
The goal here was to get a file of sentence/message pairs produced by the laird parser for the 207 sentences in the file called 209ordered.txt. This was partially achieved. A new class called LanguageTestWriter was added to the code to print out sentence/message pairs to  a file, and related code changes were made to facilitate this.
It turns out that the laird parser crashes on a number of these sentences, even though it worked on them previously.  The output of the parser on the 209ordered.txt file is called laird-209ordered.txt. This file only has 15 sentence/message pairs in it since the parser crashed on the 16th sentence. The 209-ordered-bugs.txt file has 16 sentences of the list commented out. The laird parser parses all the remaining 191 sentences without crashing, and the results are in the laird-209-ordered-bugs.txt file. The "Laird failed sentences.txt" file has a list of the 16 sentences that caused crashes, and a second copy of the list with lines from the log file for each crash.
In addition to the sentences that crashed the parser, there are 27 sentences where the parser produces a short message with a word that it considers unknown. In some cases this is correct, since the sentences is intended to define a new word. In other cases some of these words should be defined. The list of the 17 unique unknown words is in "Laird unknown-words.txt." A new smem.txt file was added that got rid of another bunch of undefined words.
The new laird-test-xbugs.config file is used to run the laird parser on a single sentences, using files in the Data/bug-test directory. The single.txt file is intended to hold a single test sentence, and the config file runs the parser to produce an output file and a log file.
Several old files that are no longer needed were deleted.