Updated the laird output for 209
Changed laird-test.config to run the laird parser on the 209ordered-bugs.txt file so that it runs all the way without crashing to give the laird-209ordered-bugs.txt output.
Also put the  disable-reference-info flag in the place where it belongs in test-lucia.soar. For now we're running with this commented out and getting all the reference-info, which we'll just ignore for now. It may be useful at some point to uncomment this.