Testing failed sentences with laird parser
Worked with John to figure out why 16 sentences from 209ordered were not working with his parser. We uncommented a number of write actions to be able to see better what's going on. This can be useful in general, so we'll keep those changes.
For the right/left wall sentences, it turns out the current parser has trouble with differernt senses of "right" and "left". We were able to tweak some code to get these sentences to work, but then we discarded those changes because it's better to have it work for the sentences that Aaron or James actually need.
A .docx file has some notes John wrote on the issues with some of these sentences. The .config file for doing this testing was renamed and changed slightly.
The general conclusion is that most of these sentences have problems because Aaron and James have changed things in smem a lot, and made the laird parser work with that. John's code that he uses has not be upgraded for all this. Therefore, for my testing of my 207 sentences, I'll just do whatever I need to do to get a target message for any sentence that the laird parser fails on. This allows me to move ahead without anyone having to spend a lot of time on trying to make old code fit new data.