Games: set up a test for blocks-world with laird
Created a new game.config file under rosie/test-agents/lucia along with files needed to run a games script under test-agents/lucia/Data/games. Ran this with the laird parser on the blocks-world.txt script in Data/games/block-puzzles, and confirmed that the parse.txt, solution.txt, and console output looked good. This will be a model for testing games scripts one at a time. Of course for now it does not work at all with the lucia parser.
Also deleted several files that are not needed any more.