Games: set up test for blocks-world with Lucia
Set up everything necessary to run a test of Lucia on the blocks-world script in the full Rosie environment. This involved setting up a test directory for this script, adding the new grammar and p-rules for sentence 001, changing the config file to run with Lucia and the right script, and saving away some results from both parsers.
At this point the agent dies somewhere on the fourth input sentence, but since we're only ready for sentence 001 that's not too surprising.
This setup will serve well for future testing as Lucia is enhanced to handle the rest of this script.