Games: got blocks-world2 working in Rosie
5/22/2020: On the Lucia side, changes were made to make Lucia correctly comprehend the blocks-world2 script. The changes to agent/lucia/act code and grammar were copied over into Rosie, and this script was run. It works properly. A set of results files are being stored away here.
One should note that the version of the script here has been changed slightly from the version gotten from James and used in the Lucia project. The Lucia version says "No." when the agent askes if it should find a solution. The version here tells Rosie to load the initial world file and find a solution. As a result, we have good solutions here.