Added print-parser-data param
This commit makes several changes that allow better control on what the laird parser prints while it is running.
Several writes in check-failure.soar were changed to be constrolled by the already existing print-parsed-messages flag, so that if the parser fails you will always know it if you have this flag on.
A new print-parser-data flag was added and used to control several writes in monitor.soar and the two smem-lexical-retrieval.soar files. When this flag is set the laird parser outputs "Next Word:" and "Retrieval Success:" messages, otherwise they are not printed.
Of course the flags must be set in some soar file.