Games: 008 parsed with Lucia test
8/20/2020: In trying to get Lucia to parse the Games 01.03 script in test mode, problems were found because the upgrade made previously was not complete. Several things had to be fixed.
1) snip-conditional.soar needed to be loaded.
2) if-structs.soar needed to be loaded.
3) if-structs-old.soar files needed to be deleted here and in the Lucia project.
4) A rule was added to initialize-lucia to set up the game-learning flag if hypothetical is true in the params.
Now Lucia runs through the whole Games 01.03 script with no crashes, and sentence 008, #4 in this script, appears to produce a correct message. There are still development issues with some of the later sentences in this script. A trace of this test run was saved under the folder for that script.