Kitchen script working with the Lucia parser!
11/21/2020: Here on the master branch, after the merge that AM did yesterday with the PL-lucia branch, and with a new add-reference-info.soar file added under agent/lucia/act/, the whole kitchen script worked properly! Amost.
The almost part is that at the end the agent did not close the fridge. It turned out this was because the definition of "closed" in smem wash changed from having a handle of closed2 to a handle of not-open1. I changed the handle to look for in the ECG, and now it all workd perfectly.