Set up for testing grounding in Lucia
4/27/2021: Lucia is not yet handling grounding for things that are not in the world but in smem. Added a kitchen agent under task-tests/grounding using files Aaron provided for testing grounding. The agent runs fine with the laird parser, but with lucia it fails on trying to ground "fork".
In this new directory are the .zip file Aaron sent and the files extracted from there.
The file he called agent.config has been renamed to kitchen.config so Git will commit it. This file has also been changed to select the lucia parser and turn the Soar debugger on. Otherwise, all is the same as what Aaron sent.
From here we can start debugging Lucia.