Added "Store the apple." to kitchen script and tested with Lucia
7/7/2021: Added a construction and a meaning schema for apple to the Lucia grammar, and added the sentence "Store the apple." to the kitchen script.
Ran this with both the laird and lucia parsers, and they produce identical test-output files which also conform to the correct-output file, which has had this sentence added to it.
With these changes the new kitchen script should work in simulation with the Lucia parser, and the added sentence makes it a bit clearer how Rosie learns a new task.