Updated Lucia to do a QUOTED-STRING lexical
8/9/2021: Made changes to next-word to present quoted strings properly to Lucia in Rosie. This works for the simple sentence "Hello Charlie". It should work for other cases as well, but that has not all been tested. These changes were carefully merged with the current state of next-word in Lucia, so now they are the same.
Then added the changes to r-nouns to add the QUOTED-STRING lexical to Rosie. Now Lucia in Rosie seems to work correctly for all three of the sentences now in the test-agents/lucia/Data/test.txt file that is changed here.
The test-agents/lucia/robot.config file was changed to make it easy to test Lucia in Rosie by itself. This is very useful at times.
Doing so also emptied out the text in test-agents/lucia/test-output.txt, but this change was discarded to keep the old version around for future reference.