Set up for testing as a standalone agent
This commit adds two _source.soar files that can be used to source a version of Rosie that can run in the Soar debugger without any Java code, just as an independent agent. The _agent_ file runs with the Laird parser, and the _lucia_ file runs with the Lucia parser.