Lucia tested OK in Rosie, RosieAgentConfigurator fixed
7/18/2022: Using the test in rosie/test-agents/task-tests/kitchen, Rosie was run correctly using the Lucia parser. To accomplish this, the parser option in the kitchen.config file was changed from laird to lucia.
In addition, some changes were needed in RosieAgentConfigurator.java. It turns out that the Java getAbsolutePath method in Windows returns a string with lots of '\' characters. Some of the code in RosieCLI doesn't like that, so after the calls to getAbsolutePath a .replaceAll("\\\\", "/") was added. Now the paths come out with forward slashes in the agent/kitchen_source.soar file, and the whole system runs just fine.